+++ src/components/Home/HeaderScene.tsx
@@ -16,7 +16,7 @@ import {
 import { getHomeScene } from '../../utils/utils';
 import { useNavigate } from 'react-router-dom-v5-compat';
 import { Home } from 'pages/Home/Home';
-import { DarkModeRocket, LightModeRocket } from './rockets';
+import { DarkModeRocket, LightModeRocket } from '../../utils/rockets';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from 'utils/analytics';
 
 export class HeaderScene extends SceneObjectBase {
